refactor(data): merge magics and earthkit presets into weather_presets.json
Both libraries are GRIB-shortName-keyed meteorological presets, so combine
them into one file and one schema instead of two vendor-named ones. The 5
shortNames earthkit-plots already overrode (2d, 2t, aod550, duaod550, tp)
are dropped from the Magics side rather than shipped as dead data.

The encoded Magics style token (magics_style, e.g. "sh_all_fM48t56i4") is
decoded once at build time into plain vmin/vmax instead of at import time,
removing _decode_magics_range/_magics_num/_MAGICS_RANGE from colors.py.
_load_magics_presets and _load_earthkit_presets collapse into a single
_load_weather_presets; _load_preset_asset goes back to handling only the
continuous cmocean case it was originally written for.

tools/build_magics_presets.py and tools/build_earthkit_presets.py are
replaced by tools/build_weather_presets.py, which fetches both sources and
merges them into one asset.

Verified byte-for-byte identical DATA_STYLES for all 101 registered
presets before and after.
