+++ src/cleopatra/data/__init__.py
@@ -1,8 +1,10 @@
 """Color presets.
 
-- ``magics_presets.json`` -- colour data and parameter labels derived from the
-  ECMWF Magics style library.
-- ``ocean_presets.json`` -- colour data.
+- ``weather_presets.json`` -- colour data and parameter labels for GRIB
+  shortName-keyed atmospheric presets, merged from ECMWF Magics and
+  earthkit-plots.
+- ``ocean_presets.json`` -- colour data for ocean/hydrology/DEM presets,
+  derived from cmocean.
 
 Kept as a package so the assets are importable via ``importlib.resources`` and
 shipped in the wheel.
